requirements.txt
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
Farama-Notifications==0.0.4
fonttools==4.55.3
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
kiwisolver==1.4.8
matplotlib==3.10.0
minigrid==3.0.0
numpy==2.2.1
packaging==24.2
pillow==11.1.0
pygame==2.6.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
